+++ NEWS.md
@@ -2,6 +2,7 @@
 ### Features
 - `check_setup()` function
 - `diagnostics()` function
+- `rho2logsdr` transformations
 
 ## version 1.0.3
 Compatibility update for the rstan 2.31.